Don't add blank lines after comments in `inherit`
Fixes #376

Note that handling of extra newlines in `inherit` statements is still a
little janky; usually blank lines will be removed from formatted output,
but in some cases (e.g. between an end-of-line comment and a full-line
comment) they will be preserved. See the tests added for a few examples
of this.